+++ spring-boot-starter/spring-boot-starter-web/src/test/java/com/getyourguide/openapi/validation/integration/controller/DefaultRestController.java
@@ -1,10 +1,10 @@
 package com.getyourguide.openapi.validation.integration.controller;
 
+import com.getyourguide.openapi.validation.test.exception.WithResponseStatusException;
+import com.getyourguide.openapi.validation.test.exception.WithoutResponseStatusException;
 import com.getyourguide.openapi.validation.test.openapi.web.DefaultApi;
 import com.getyourguide.openapi.validation.test.openapi.web.model.PostTestRequest;
 import com.getyourguide.openapi.validation.test.openapi.web.model.TestResponse;
-import com.getyourguide.openapi.validation.integration.exception.WithResponseStatusException;
-import com.getyourguide.openapi.validation.integration.exception.WithoutResponseStatusException;
 import java.time.LocalDate;
 import java.util.Objects;
 import org.springframework.http.ResponseEntity;
