+++ test/test-utils/src/main/java/com/getyourguide/openapi/validation/test/exception/WithoutResponseStatusException.java
@@ -1,4 +1,4 @@
-package com.getyourguide.openapi.validation.integration.exception;
+package com.getyourguide.openapi.validation.test.exception;
 
 public class WithoutResponseStatusException extends RuntimeException {
